% tinyrox says don't edit this manually, but it can't stop you!
\name{.basic_auth}
\alias{.basic_auth}
\title{HTTP Basic authorization header value for client credentials}
\usage{
.basic_auth(id, secret)
}
\description{
HTTP Basic authorization header value for client credentials
}
\keyword{internal}
