% tinyrox says don't edit this manually, but it can't stop you!
\name{.codex_device_start}
\alias{.codex_device_start}
\title{Start the device-authorization flow: get a user code to display}
\usage{
.codex_device_start(client, post = .codex_post_json)
}
\arguments{
\item{post}{JSON-POST function, injectable for testing.}
}
\description{
Start the device-authorization flow: get a user code to display
}
\keyword{internal}
