% tinyrox says don't edit this manually, but it can't stop you!
\name{.codex_exchange}
\alias{.codex_exchange}
\title{Exchange the device authorization code (with its PKCE verifier) for a token}
\usage{
.codex_exchange(client, code, verifier)
}
\description{
Exchange the device authorization code (with its PKCE verifier) for a token
}
\keyword{internal}
