% tinyrox says don't edit this manually, but it can't stop you!
\name{.codex_finalize}
\alias{.codex_finalize}
\title{Attach the ChatGPT account id (from the access-token JWT) to a token}
\usage{
.codex_finalize(token)
}
\description{
Attach the ChatGPT account id (from the access-token JWT) to a token
}
\keyword{internal}
