% tinyrox says don't edit this manually, but it can't stop you!
\name{.codex_post_json}
\alias{.codex_post_json}
\title{POST a JSON body and parse the JSON response}
\usage{
.codex_post_json(url, body)
}
\description{
The device endpoints take JSON (the token endpoint takes form-encoding, which
is what [.token_request] handles).
}
\keyword{internal}
