% tinyrox says don't edit this manually, but it can't stop you!
\name{.form_encode}
\alias{.form_encode}
\title{Form-encode a named list as application/x-www-form-urlencoded, dropping NULLs}
\usage{
.form_encode(fields)
}
\description{
Form-encode a named list as application/x-www-form-urlencoded, dropping NULLs
}
\keyword{internal}
