% tinyrox says don't edit this manually, but it can't stop you!
\name{.parse_request_query}
\alias{.parse_request_query}
\title{Parse the query parameters from an HTTP request line}
\usage{
.parse_request_query(req_line)
}
\description{
Parse the query parameters from an HTTP request line
}
\keyword{internal}
