% tinyrox says don't edit this manually, but it can't stop you!
\name{.token_request}
\alias{.token_request}
\title{POST a grant to the token endpoint and parse the result}
\usage{
.token_request(client, fields)
}
\description{
POST a grant to the token endpoint and parse the result
}
\keyword{internal}
