% tinyrox says don't edit this manually, but it can't stop you!
\name{oauth_cache_path}
\alias{oauth_cache_path}
\title{Default on-disk cache path for a client's token}
\usage{
oauth_cache_path(client)
}
\arguments{
\item{client}{A [oauth_client].}
}
\value{
Path to the token cache file under \code{tools::R_user_dir}.
}
\description{
Default on-disk cache path for a client's token
}
