% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{tinyoauth-package}
\alias{tinyoauth}
\alias{tinyoauth-package}
\title{\packageTitle{tinyoauth}}
\description{\packageDescription{tinyoauth}}
\section{Package Content}{\packageIndices{tinyoauth}}
\author{\packageAuthor{tinyoauth}}
\section{Maintainer}{\packageMaintainer{tinyoauth}}
\keyword{internal}
